Move UserMenu to its own directory.
